requirements.txt
PyYAML
StrEnum
diskcache
jupyterlab
mineru[core]
ollama
pypdf
tiktoken
xxhash
torch
doclayout_yolo
ultralytics
transformers
ftfy